SAW-512: Accept multiple artifacts to prevent testing errors
Previously only a single artifact was accepted and an error thrown
otherwise.  This turns out to be an issue during testing when multiple
artifacts might be provided in a request.  So silently accept the rest
of the artifacts for now and add a note to implement support for
multiple artifacts later.