SAW-870: Take runtime filters from analysis in request body
Previously the runtime filters were taken from a special property
provided with the execute action.  Change it so that the runtime
filters instead are taken from the analysis object provided in the
request object.  This corresponds better to how the runtime filters
were originally planned to be provided to the back-end.