SAW-1309: Fix executing manually edited query
Previously a query was always built from the analysis model, even
though a manual query had been entered.  If no column was checked,
this lead to an "at least one column must be checked" error.  Change
it to not attempt building a query from the analysis model if a manual
query is provided.