SAW-1520: Remove the jvm-exit-on-fatal-error false setting
After the changes to result handling to address the out of memory
errors, there should no longer be a need for this workaround.  Setting
jvm-exit-on-fatal-error to false has the risk of freezing the entire
JVM, so it is better to leave it to its default of true and let the
JVM be restarted by system services into a fresh working state.