SAW-1523: Commit consumer offsets explicitly before closing
The MapR streams documentation says auto commit of consumer offsets is
enabled by default.  However, it was discovered that the consumer was
getting the same messages over again, indicating the offset was not
getting committed.  So commit the consumer offset explicitly before
closing as a workaround to ensure the offsets get committed.