SAW-1188: Update path for accessing SAW Web root in test
Previously the root path ("/") was redirected to the SAW Web
root ("/web").  Now however, in local deployments, the root serves a
starting page for accessing services in the local deployments.  So
change the URL used for testing the SAW Web root to use the full
path ("/web") instead.  In production deployments the root is still
redirected to the SAW Web root.