SAW-1345: Move external documentation into rendering pipeline
Move documentation that is published for external consumption (similar
to the release package) under "saw-dist/src/main/asciidoc" and include
it in the documentation rendering pipeline.  Remove the SAW User Guide
placeholder as it is now maintained elsewhere.

The internal developer documentation will remain in the top-level
"doc" directory, as developers read it as plain text/Markdown and it
doesn't need to be rendered.