SAW-1823: Add and check error indicator to analysis result node
While testing joins between tables where the join column is selected
in both tables, it was noticed that an error in the execution would
not be immediately detected by the calling Transport Service, instead
timing out waiting on the results.  Add an error indicator as a first
workaround to allow the Transport Service to detect the error
condition faster.