SAW-1415: Generate database password on target host instead
Previously the password was generated and stored on the deploy host
where Ansible is invoked from.  This lead to the password file being
stored with the user that Ansible was being invoked as the owner.
Change it to be stored on the target host where MariaDB is running
using root as the owner instead.