SAW-1823: Ensure Transport Service detects when execution completes
The previous approach to letting the Transport Service know when an
execution had completed by looking at the result node was impractical,
as a result node might not be created due to execution errors for
example.  So replace that approach by instead using MapR-FS file
markers to get more reliable reporting of when an execution has
completed.