SAW-2975: Make integration tests use Docker Machine by default
Currently a local Docker engine is used for integration tests.  This
has however the downside that the environment is limited to the
resources available on the developer's local laptop, or the continuous
integration agent.  Therefore change to using a remote Docker Machine
which can be adjusted to meet memory needs.

The option of using a local Docker engine is still available and might
be faster for those who have enough memory available locally.