SIP-1128: Document health checks and high availability
Add documentation on the existence and location of the services health
checks to the Operations and Design Guides.  The health checks might
be used by operations to monitor the health of individual services.

Also describe the purpose of health checks and approach to high
availability for stateless services in the Design Guide.