SIP-3542: Remove dependency on saw-deploy from saw-logger
The saw-logger service is now installed in the base image used by all
nodes.  Only the admin node actually has the saw-deploy service,
because deployment is started from there, so there is no need to have
saw-logger depend on saw-deploy everywhere.  It will anyway be started
early in the boot and should include most of saw-deploy logging.