SIP-3542: Add forward proxy for accessing internal resources
Add Squid as a forward proxy to development deployments to enable
accessing internal nodes using their own internal hostnames and ports.
This can be used for example to access the Spark master and worker web
interfaces, without having to rewrite URLs.  This approach works
better with cloud deployments, where it might be difficult to open up
a larger number of ports in AWS security groups.

Also rename the "start" directory to "dev" to indicate it contains
resources for development.