SIP-3539: Use "revision.nodash" as RPM version identifier
Previously some services defaulted to the Maven project version, which
can contain dashes.  Dashes should not be part of RPM versions.  This
seems to have caused issues with Yum and Ansible complaining when
trying to install a package of the same version.  Update services to
use the nodash version of the Maven project version.