SIP-3539: Use the same 20 minute timeout for all containers
Previously the admin node had a timeout of 20 minutes (to account for
the deployment of SIP), while other containers only had a timeout of 2
minutes.  It turns out that in local deployment, the MapR container
can sometimes take more than 2 minutes to come up.  So change all
containers to use the same 20 minute timeout, which also simplifies
the configuration.