SIP-3317: Make upgrade testing work in continuous integration
Previously the upgrade testing was triggered for "mvn
-Ddocker-start=cloud" which sets a custom Maven profile "defaultGoal",
but not "mvn package" that relies on Maven plug-ins registering goals
for phases.  This caused the upgrade testing flag to not work in
continuous integration.  So add a Maven plug-in execution to ensure
that the upgrade testing also works (if enabled by
"-Dsip.upgrade.skip=false") for continuous integration builds.