SIP-3317: Handle race condition in adding category
The add category controller logic contains a race condition between
checking for category existence and actually creating it if missing.
If the code is executed simultaneously on multiple instances of the
service, it is possible for both to attempt adding a new category.
Add "ON DUPLICATE KEY UPDATE" to handle this transparently.

Additionally, if MariaDB cluster replication is used, simulatenous
inserts using the same primary key might result in a deadlock error,
in which case the statement needs to be retried.