SIP-4969: Add documentation on adjusting RTPS stream rates
By default the Spark Streaming connector to MapR Streams will pull in
as many records as it can.  If the RTPS has been down and a backlog
has been built up, this can lead to the RTPS Spark application running
out of memory while trying to process a large amount of records in a
single batch.  So add documentation on how to limit the number of
records read from MapR Streams for any single batch, as a way of
limiting memory usage.