SIP-5376: Add check to ensure sip-rtis exists in configuration
Ensure that SIP environment configuration file contains the "sip-rtis"
role, as is required for the HAProxy configuration to render properly.
It can however be left empty if RTIS is not going to be installed.
Previous versions of SIP did not have the "sip-rtis" role, so older
environments might not have this role in their configuration files.