SIP-5433: Do not embed MapR dependencies in SIP package
They are supposed to be provided by the MapR installation instead of
being bundled with the application.  This ensures the right version of
the libraries is used and avoids version mismatch errors.

Instead add a utility "sip-mapr-classpath" that prepare a MapR
classpath suitable for use with the Spring Boot PropertiesLauncher and
pass it using the "loader.path" option.