SIP-2286: Add retry of stopping in service handlers
Occasionally deployments will fail with "Job for service canceled".
Preferably the root cause for this would be found and fixed, but until
then add a retry to recover and proceed with the deployment.

This same workaround was already applied to a number of other
services, and it seems to help recovering from those failures, so add
the workaround to the rest of the services too.