SIP-4612: Add environment parameter for allowed origins
The new SIP environment parameter allows configuring SIP development
and testing environments to accept requests to APIs from other origins
than the host on which the SIP environment is running (for example
from localhost on the machine that a developer is running the SIP
front-end on locally).  It also allows specifying third-party hosts
that integrate with SIP APIs as allowed origins.