SIP-3317: Ensure saw-security-arbitrator role exists
Ensure that SIP environment configuration file contains the expected
roles, as they are required for certain configurations to render
properly.  They can however be left empty with no hosts if that role
is not going to be installed.  Previous versions of SIP did not have
all current roles, so older environments might not have all of these
roles in their configuration files.