SIP-6043: Increase the default cloud timeout to eight hours
Most developers seem to need their machine for the full workday, so
increase the default timeout to eight hours.  This avoids losing the
cloud environment in the middle of the workday.  If the costs turn out
to be too high from this change, the default timeout will have to be
reverted back to four hours again.