SIP-5967: Exclude JUnit from spring-boot-maven-plugin packaging
The spring-boot-starter-test plug-in includes JUnit as a dependency,
but puts it in the "compile" scope instead of the "test" scope.  This
causes the JUnit libraries to be included in production packaging.
This further causes problems for components that are designed to only
be loaded at testing or run time, based on the existence of unit
testing libraries.  As a workaround, explicitly exclude JUnit from the
final package.