SIP-7163: Expose incoming HTTPS on separate port 443
Previously both HTTP and HTTPS were exposed on port 80 depending on
the "sip_secure" configuration parameter.  However, if
docker-maven-plugin is configured in "dist/pom.xml" to map multiple
ports to port 80, only one of them seems to take effect.  So as a
workaround expose HTTPS on a separate port, so that there will be a
one-to-one mapping of all ports exposed by Docker.